refactoring train code.
